fix: Fix random leader intro quotes
For whatever reason, changing the IntroductionQuote property for each leader wasn't correctly being applied; the intro text was still being fetched from the database as if it hadn't been changed. The new approach is to add each intro text per leader. It's a lot more work but it seems to work, and doesn't require changing the Civilizations table.

Also update the readme to reflect the new changes