feat: Disable logic for showing victory prologue popups
In the end it's too messy to implement and complicated to test. Affinity victories could be programmed to only show for the highest affinity, but then what about contact or domination victory? The logic for tracking victory progress is split into two places and should be combined. Testing it all is messy as there can be tied affinity levels. I think what put me over the top was worrying about adding logic to attempt to detect if the victories had been modded. In the end it's too much effort for too little value.