feat: Re-scale flavours to a max of 9
It seems that the max flavour value is 10 but the game can randomise it within 2 fewer or additional points. So if it's set to 12, no matter if the game adds or subtracts 2, it will always end up at 10 (the max value). By setting it to 9, in practice it means 7-10. So this gives a strong inclination but with a small amount of room for variation.

Also add some commented code from an attempt to influence leaders to pick techs that would increase their affinity